доопрацювання
