--- controllers/auths.js
@@ -1,150 +1,0 @@
-const { User } = require("../models/user");
-const { ctrlWrapper, HttpError, sendEmail } = require("../helpers");
-const bcrypt = require("bcrypt");
-const gravatar = require("gravatar");
-const path = require("path");
-const fs = require("fs/promises");
-const Jimp = require("jimp");
-const { nanoid } = require("nanoid");
-const jwt = require("jsonwebtoken");
-const { SECRET_KEY, BASE_URL } = process.env;
-const avatarDir = path.join(__dirname, "../", "public", "avatars");
-
-const register = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  const avatarURL = gravatar.url(email);
-  if (user) {
-    throw HttpError(409, "Email in use");
-  }
-  const hashPassword = await bcrypt.hash(password, 10);
-  const verificationToken = nanoid();
-  const newUser = await User.create({
-    ...req.body,
-    password: hashPassword,
-    avatarURL,
-    verificationToken,
-  });
-  const verifyEmail = {
-    to: email,
-    subject: "Verify email",
-    html: `<a target="_blank" href="${BASE_URL}/users/verify/${verificationToken}">Click verify email</a>`,
-  };
-  await sendEmail(verifyEmail);
-  res.status(201).json({
-    user: {
-      email: newUser.email,
-      subscription: newUser.subscription,
-    },
-  });
-};
-
-const verifyEmail = async (req, res) => {
-  const { verificationToken } = req.params;
-  const user = await User.findOne({ verificationToken });
-  if (!user) {
-    throw HttpError(404, "User not found");
-  }
-  await User.findByIdAndUpdate(user._id, {
-    verify: true,
-    verificationToken: "",
-  });
-  res.json({ message: "Verification successful" });
-};
-
-const resendVerifyEmail = async (req, res) => {
-  const { email } = req.body;
-  const user = await User.findOne({ email });
-  if (!user) {
-    throw HttpError(404, "Email not found");
-  }
-  if (user.verify) {
-    throw HttpError(400, "Verification has already been passed");
-  }
-  const verifyEmail = {
-    to: email,
-    subject: "Verify email",
-    html: `<a target="_blank" href="${BASE_URL}/users/verify/${user.verificationToken}">Click verify email</a>`,
-  };
-  await sendEmail(verifyEmail);
-  res.json({
-    message: "Verification email sent",
-  });
-};
-
-const login = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  if (!user) {
-    throw HttpError(401, "Email or password is wrong");
-  }
-  if (!user.verify) {
-    throw HttpError(401, "Email not verified");
-  }
-  const comparePassword = await bcrypt.compare(password, user.password);
-  if (!comparePassword) {
-    throw HttpError(401, "Email or password is wrong");
-  }
-  const payload = {
-    id: user._id,
-  };
-  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "23h" });
-  await User.findByIdAndUpdate(user._id, { token });
-  res.status(200).json({
-    token,
-    user: {
-      email: user.email,
-      subscription: user.subscription,
-    },
-  });
-};
-
-const logout = async (req, res) => {
-  const { _id } = req.user;
-  await User.findByIdAndUpdate(_id, { token: "" });
-  res.status(204).json({ message: "Logout success" });
-};
-
-const current = async (req, res) => {
-  const { email, subscription } = req.user;
-  res.json({ email, subscription });
-};
-
-const subscription = async (req, res) => {
-  const { _id } = req.user;
-  const result = await User.findByIdAndUpdate(_id, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.json({
-    email: result.email,
-    subscription: result.subscription,
-  });
-};
-
-const updateAvatar = async (req, res) => {
-  const { _id } = req.user;
-  const { path: tempUpload, originalname } = req.file;
-  const fileName = `${_id}_${originalname}`;
-  const resultUpload = path.join(avatarDir, fileName);
-  await fs.rename(tempUpload, resultUpload);
-  const avatarURL = path.join("avatar", fileName);
-  const image = await Jimp.read(resultUpload);
-  image.resize(250, 250);
-  await image.writeAsync(resultUpload);
-  await User.findByIdAndUpdate(_id, { avatarURL });
-  res.json({ avatarURL });
-};
-
-module.exports = {
-  register: ctrlWrapper(register),
-  verifyEmail: ctrlWrapper(verifyEmail),
-  resendVerifyEmail: ctrlWrapper(resendVerifyEmail),
-  login: ctrlWrapper(login),
-  logout: ctrlWrapper(logout),
-  current: ctrlWrapper(current),
-  subscription: ctrlWrapper(subscription),
-  updateAvatar: ctrlWrapper(updateAvatar),
-};
