+++ controllers/auths/current.js
@@ -1,0 +1,6 @@
+const current = async (req, res) => {
+  const { email, subscription } = req.user;
+  res.json({ email, subscription });
+};
+
+module.exports = current;
