+++ controllers/auths/login.js
@@ -1,0 +1,34 @@
+const jwt = require("jsonwebtoken");
+const bcrypt = require("bcrypt");
+const { HttpError } = require("../../helpers");
+const { User } = require("../../models/user");
+const { SECRET_KEY } = process.env;
+
+const login = async (req, res) => {
+  const { email, password } = req.body;
+  const user = await User.findOne({ email });
+  if (!user) {
+    throw HttpError(401, "Email or password is wrong");
+  }
+  if (!user.verify) {
+    throw HttpError(401, "Email not verified");
+  }
+  const comparePassword = await bcrypt.compare(password, user.password);
+  if (!comparePassword) {
+    throw HttpError(401, "Email or password is wrong");
+  }
+  const payload = {
+    id: user._id,
+  };
+  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "23h" });
+  await User.findByIdAndUpdate(user._id, { token });
+  res.status(200).json({
+    token,
+    user: {
+      email: user.email,
+      subscription: user.subscription,
+    },
+  });
+};
+
+module.exports = login;
