+++ controllers/auths/logout.js
@@ -1,0 +1,8 @@
+const { User } = require("../../models/user");
+
+const logout = async (req, res) => {
+  const { _id } = req.user;
+  await User.findByIdAndUpdate(_id, { token: "" });
+  res.status(204).json({ message: "Logout success" });
+};
+module.exports = logout;
