+++ controllers/auths/updateAvatar.js
@@ -1,0 +1,21 @@
+const path = require("path");
+const fs = require("fs/promises");
+const Jimp = require("jimp");
+const { User } = require("../../models/user");
+const avatarDir = path.join(__dirname, "../", "../", "public", "avatars");
+
+const updateAvatar = async (req, res) => {
+  const { _id } = req.user;
+  const { path: tempUpload, originalname } = req.file;
+  const fileName = `${_id}_${originalname}`;
+  const resultUpload = path.join(avatarDir, fileName);
+  await fs.rename(tempUpload, resultUpload);
+  const avatarURL = path.join("avatar", fileName);
+  const image = await Jimp.read(resultUpload);
+  image.resize(250, 250);
+  await image.writeAsync(resultUpload);
+  await User.findByIdAndUpdate(_id, { avatarURL });
+  res.json({ avatarURL });
+};
+
+module.exports = updateAvatar;
