+++ controllers/contacts.js
@@ -48,6 +48,7 @@ const putContact = async (req, res) => {
 
 const patchContacts = async (req, res) => {
   const { contactId } = req.params;
+  console.log(contactId);
   const result = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
   });
